[#509] merged notifications into landscape context, created util for FE tests
