+++ src/main/app/src/Components/Notification/Notification.tsx
@@ -1,70 +1,35 @@
-import React, { useState, useEffect, useContext, ReactElement } from 'react';
-import { INotificationMessage } from "../../interfaces";
-import { Client, StompSubscription } from '@stomp/stompjs';
-import { withBasePath } from '../../utils/API/BasePath';
+import React, { ReactElement, useContext, useEffect, useState } from 'react';
 import IconButton from '@material-ui/core/IconButton';
-import { Badge} from '@material-ui/core';
-import { DynamicFeed } from "@material-ui/icons";
-import { NotificationContext } from '../../Context/NotificationContext';
+import { Badge } from '@material-ui/core';
+import { DynamicFeed } from '@material-ui/icons';
 import Changes from './Changes';
-import componentStyles from "../../Resources/styling/ComponentStyles";
+import componentStyles from '../../Resources/styling/ComponentStyles';
+import { LandscapeContext } from '../../Context/LandscapeContext';
 
 interface Props {
   setSidebarContent: Function;
 }
 
 /**
- * Logic component for notifications. Subscribes via websockets to server events.
+ * Displaying server side notifications.
  *
  *
  */
 const Notification: React.FC<Props> = ({ setSidebarContent }) => {
 
   const classes = componentStyles();
-  const backendUrl = withBasePath('/subscribe');
-  const protocol = window.location.protocol !== 'https:' ? 'ws' : 'wss';
-
-  const [socketUrl] = useState(protocol + `://${backendUrl.replace(/^https?:\/\//i, '')}`);
-  const [subscriptions, setSubscriptions] = useState<StompSubscription[]>([]);
   const [newChanges, setNewChanges] = useState<Boolean>(false);
   const [renderedChanges, setRenderedChanges] = useState<ReactElement | null>(null);
-  const notificationContext = useContext(NotificationContext);
-
-  const [client] = useState(
-    new Client({
-      brokerURL: socketUrl,
-      onConnect: () => {
-        const subscriptions: StompSubscription[] = [];
-        const eventSubscription = client.subscribe('/topic/events', (message) => {
-          const notificationMessage: INotificationMessage = JSON.parse(message.body);
-          if (notificationMessage.type === 'ProcessingFinishedEvent') {
-            notificationContext.next(notificationMessage);
-            setNewChanges(true);
-          }
-        });
-
-        subscriptions.push(eventSubscription);
-        setSubscriptions(subscriptions);
-      },
-    })
-  );
-
-  useEffect(() => {
-    client.activate();
-    return () => {
-      subscriptions.forEach((subscription) => {
-        subscription.unsubscribe();
-      });
-    };
-  }, [client, subscriptions]);
+  const landscapeContext = useContext(LandscapeContext);
+  
 
   /**
    * render changes,
    */
   useEffect(() => {
-    if (notificationContext.notification == null) return;
-    setRenderedChanges(<Changes notification={notificationContext.notification} />);
-  }, [notificationContext.notification]);
+    if (landscapeContext.notification == null) return;
+    setRenderedChanges(<Changes notification={landscapeContext.notification} />);
+  }, [landscapeContext.notification]);
 
   return (
     <Badge color='secondary' variant='dot' overlap='circle' invisible={!newChanges} title={'Recent changes'}>
