+++ src/main/app/src/Context/LandscapeContext.tsx
@@ -1,11 +1,14 @@
 import React, { useEffect, useState } from 'react';
-import { IAssessment, IAssessmentProps, ILandscape } from '../interfaces';
+import { IAssessment, IAssessmentProps, ILandscape, INotificationMessage } from '../interfaces';
 import { get } from '../utils/API/APIClient';
+import { withBasePath } from '../utils/API/BasePath';
+import { Client, StompSubscription } from '@stomp/stompjs';
 
 export interface LandscapeContextType {
   readonly landscape: ILandscape | null;
   readonly assessment: IAssessment | null;
   readonly identifier: string | null;
+  readonly notification: INotificationMessage | null;
   next: (identifier: string | null) => void;
   getAssessmentSummary: (fqi: string) => IAssessmentProps | null;
 }
@@ -14,6 +17,7 @@ export const LandscapeContext = React.createContext<LandscapeContextType>({
   landscape: null,
   assessment: null,
   identifier: null,
+  notification: null,
   next: () => {},
   getAssessmentSummary: () => {
     return null;
@@ -23,7 +27,8 @@ export const LandscapeContext = React.createContext<LandscapeContextType>({
 /**
  * Provides the current landscape and assessment.
  *
- * Responsible for loading objects from the API.
+ * - Responsible for loading objects from the API.
+ * - Responsible to fetch change events (notification, subscribes via websockets to server events)
  *
  * @param props
  * @constructor
@@ -31,9 +36,51 @@ export const LandscapeContext = React.createContext<LandscapeContextType>({
 const LandscapeContextProvider: React.FC<{}> = (props) => {
   const [landscape, setLandscape] = useState<ILandscape | null>(null);
   const [assessment, setAssessment] = useState<IAssessment | null>(null);
+  const [notification, setNotification] = useState<INotificationMessage | null>(null);
   const [identifier, setIdentifier] = useState<string | null>(null);
 
-  //load landscape
+  const backendUrl = withBasePath('/subscribe');
+  const protocol = window.location.protocol !== 'https:' ? 'ws' : 'wss';
+
+  const [socketUrl] = useState(protocol + `://${backendUrl.replace(/^https?:\/\//i, '')}`);
+  const [subscriptions, setSubscriptions] = useState<StompSubscription[]>([]);
+
+  /**
+   * Subscribe to event stream
+   */
+  const [client] = useState(
+    new Client({
+      brokerURL: socketUrl,
+      onConnect: () => {
+        const subscriptions: StompSubscription[] = [];
+        const eventSubscription = client.subscribe('/topic/events', (message) => {
+          const notificationMessage: INotificationMessage = JSON.parse(message.body);
+          if (notificationMessage.type === 'ProcessingFinishedEvent') {
+            setNotification(notificationMessage);
+          }
+        });
+
+        subscriptions.push(eventSubscription);
+        setSubscriptions(subscriptions);
+      },
+    })
+  );
+
+  /**
+   * Activate client
+   */
+  useEffect(() => {
+    client.activate();
+    return () => {
+      subscriptions.forEach((subscription) => {
+        subscription.unsubscribe();
+      });
+    };
+  }, [client, subscriptions]);
+
+  /**
+   * Load the landscape and assessment data when the identifier changes.
+   */
   useEffect(() => {
     if (identifier == null) {
       console.debug(`Identifier not present`);
@@ -56,6 +103,7 @@ const LandscapeContextProvider: React.FC<{}> = (props) => {
         landscape: landscape,
         assessment: assessment,
         identifier: identifier,
+        notification: notification,
         next: (nextId) => {
           if (identifier === nextId) return;
           console.debug('New identifier', nextId);
